+++ silo-oracles/contracts/interfaces/IOracleScalerFactory.sol
@@ -8,8 +8,10 @@ interface IOracleScalerFactory {
 
     /// @notice Create a new oracle scaler
     /// @param _quoteToken The quote token for this oracle to support.
+    /// @param _externalSalt The external salt to be used together with factory salt
     /// @return oracleScaler The oracle scaler created
     function createOracleScaler(
-        address _quoteToken
+        address _quoteToken,
+        bytes32 _externalSalt
     ) external returns (ISiloOracle oracleScaler);
 }